build(korm): polymorphic ddl for columns, IdTable abstraction, clickhouse integration

+++ lib/src/main/kotlin/com/rmp/lib/utils/korm/TableRegister.kt
@@ -7,14 +7,14 @@ import com.rmp.lib.utils.korm.query.QueryParseData
 import com.rmp.lib.utils.korm.query.QueryType
 
 object TableRegister {
-    val tables = mutableMapOf<String, Table>()
+    val tables = mutableMapOf<String, Pair<DbType, Table>>()
 
-    fun register(vararg tables: Table) {
-        this.tables += tables.associateBy { it.tableName_ }
+    fun register(dbType: DbType, vararg tables: Table) {
+        this.tables += tables.associate { it.tableName_ to Pair(dbType, it) }
     }
 
     fun findColumns(tableName: String, columns: List<String>): List<Column<*>> {
-        val table = tables[tableName] ?: throw Exception("Unknown model $tableName")
+        val (_, table) = tables[tableName] ?: throw Exception("Unknown model $tableName")
 
         return columns.mapNotNull {
             table.columns[it]
@@ -25,46 +25,52 @@ object TableRegister {
         queryParseData.map { (tableName, columns) -> findColumns(tableName, columns) }.flatten()
 }
 
-fun Table.initTable(modifier: StringBuilder.() -> Unit = {}, forceRecreate: Boolean = false): QueryDto {
+fun Table.initTable(dbType: DbType, forceRecreate: Boolean = false): QueryDto {
     val builder = object : QueryBuilder(this) {}.apply {
         setQuery(StringBuilder().apply {
-            if (forceRecreate) {
+            // Disable for clickhouse due to its incapability to multi statement execution
+            if (forceRecreate && dbType != DbType.CLICKHOUSE) {
                 append("drop table if exists ")
                 append(tableName_)
-                append(" cascade;")
+                if (this@initTable is IdTable)
+                    append(" cascade;")
+                else
+                    append(";")
             }
-            append("create table ")
+            append("create table if not exists ")
             append(tableName_)
             append(" (")
 
             append(
-                columns.values.joinToString(separator = ",") { it.ddl() }
+                columns.values.joinToString(separator = ",") { it.ddl(dbType) }
             )
 
             append(",primary key (")
             append(columns.values.filter { it.type.isPk }.joinToString(separator = ",") { it.name })
             append(")")
 
-            if (references.isNotEmpty()) {
-                references.forEach { target, refs ->
-                    refs.forEach { ref ->
-                        append(",foreign key(")
-                        append(ref.sourceColumn.name)
-                        append(") references ")
-                        append(target.tableName_)
-                        append("(")
-                        append(target.id.name)
-                        append(")")
-                        append(" on delete ")
-                        append(ref.deleteOption.sql)
-                        append(" on update ")
-                        append(ref.updateOption.sql)
+            if (this@initTable is IdTable) {
+                if (references.isNotEmpty()) {
+                    references.forEach { target, refs ->
+                        refs.forEach { ref ->
+                            append(",foreign key(")
+                            append(ref.sourceColumn.name)
+                            append(") references ")
+                            append(target.tableName_)
+                            append("(")
+                            append(target.id.name)
+                            append(")")
+                            append(" on delete ")
+                            append(ref.deleteOption.sql)
+                            append(" on update ")
+                            append(ref.updateOption.sql)
+                        }
                     }
                 }
             }
 
             append(");")
-        }.apply(modifier))
+        })
     }
 
     return QueryDto.executeQuery(QueryType.DDL, builder.getQuery())
